moved ProjectName const to vscode-extension
