+++ apps/vs-code-react/src/run-service/types.ts
@@ -30,15 +30,6 @@ export interface RunDisplayItem {
   status: string;
 }
 
-export const ProjectName = {
-  export: 'export',
-  overview: 'overview',
-  review: 'review',
-  designer: 'designer',
-  dataMapper: 'dataMapper',
-} as const;
-export type ProjectNameType = (typeof ProjectName)[keyof typeof ProjectName];
-
 export interface Workflow {
   id: string;
   name: string;
