+++ libs/vscode-extension/src/lib/models/project.ts
@@ -6,6 +6,15 @@ import type { WorkflowProjectType } from './workflow';
 import type { IActionContext } from '@microsoft/vscode-azext-utils';
 import type { Uri, WorkspaceFolder } from 'vscode';
 
+export const ProjectName = {
+  export: 'export',
+  overview: 'overview',
+  review: 'review',
+  designer: 'designer',
+  dataMapper: 'dataMapper',
+} as const;
+export type ProjectNameType = (typeof ProjectName)[keyof typeof ProjectName];
+
 export const ProjectSource = {
   Remote: 'Remote',
   Local: 'Local',
